+++ apps/analog-app/src/stories/button.component.ts
@@ -2,12 +2,12 @@ import { CommonModule } from '@angular/common';
 import { Component, Input, Output, EventEmitter } from '@angular/core';
 
 @Component({
-  selector: 'storybook-button',
+  selector: 'analogjs-storybook-button',
   standalone: true,
   imports: [CommonModule],
   template: ` <button
     type="button"
-    (click)="onClick.emit($event)"
+    (click)="clicked.emit($event)"
     [ngClass]="classes"
     [ngStyle]="{ 'background-color': backgroundColor }"
   >
@@ -46,7 +46,7 @@ export class ButtonComponent {
    * Optional click handler
    */
   @Output()
-  onClick = new EventEmitter<Event>();
+  clicked = new EventEmitter<Event>();
 
   public get classes(): string[] {
     const mode = this.primary