+++ packages/trpc/src/lib/client/client.ts
@@ -14,7 +14,7 @@ import {
   tRPC_CACHE_STATE,
 } from './cache-state';
 import { createTRPCRxJSProxyClient } from './trpc-rxjs-proxy';
-// @ts-ignore
+// @ts-expect-error The internal FetchEsque type is not exported from the public API.
 import { FetchEsque } from '@trpc/client/dist/internals/types';
 
 export type TrpcOptions<T extends AnyRouter> = {